Move token resolution functions into their own file.
